Made all file decrypting, patching, and etc now work within temporary directories to reduce clutter and leftover files if something went wrong.
Also got rid of unnecessary json files as they were for me when I just started making this project, which was actually all based in Nano :P

I tried to get rid of shell but just a few of the binaries I need to use are refusing to work unless I'm using shell, so fml.

Code is still trash but at least we're working within temporary directories which I was wanting to add.