Rewrote the entire project.
I first made this in Nano, so a lot of it was bad.

I hardcoded a lot and that was janky af, but I was going to update eventually.

That promise is finally here!

The entire project is MUCH more dynamic and does less janky things.

Also, I encrypted everything that gets patched. Before I left all of them decrypted.

LLB needs to be encrypted for 24Kpwn/Alloc8 to work anyway.

I haven't tested restoring yet, but I will do so soon.

This is the biggest update to this project in terms of overall quality.